+++ quartermaster/src/main/java/com/github/reflectoring/haljson/HalJsonController.java
@@ -1,6 +1,5 @@
 package com.github.reflectoring.haljson;
 
-import com.github.reflectoring.Link;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
