+++ quartermaster/src/main/java/com/github/reflectoring/haljson/HalJsonResource.java
@@ -4,7 +4,6 @@ import com.fasterxml.jackson.annotation.JsonAnyGetter;
 import com.fasterxml.jackson.annotation.JsonAnySetter;
 import com.fasterxml.jackson.annotation.JsonInclude;
 import com.fasterxml.jackson.annotation.JsonProperty;
-import com.github.reflectoring.Link;
 
 import java.util.ArrayList;
 import java.util.HashMap;
