+++ quartermaster/src/main/java/com/github/reflectoring/haljson/Link.java
@@ -1,4 +1,4 @@
-package com.github.reflectoring;
+package com.github.reflectoring.haljson;
 
 import com.fasterxml.jackson.annotation.JsonIgnore;
