+++ quartermaster/src/main/java/com/github/reflectoring/infiniboard/quartermaster/Quartermaster.java
@@ -1,4 +1,4 @@
-package com.github.reflectoring;
+package com.github.reflectoring.infiniboard.quartermaster;
 
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 import org.springframework.boot.builder.SpringApplicationBuilder;
